odla trt runs complete

+++ lib/transforms/type_legalizer.cc
@@ -1370,7 +1370,7 @@ static void RunOnInstruction(UniqueInst* inst) {
   // FIXME: result 0 has at most the same number of
   // elements of the input
   inst->GetResultsTypes()[0] = type0;
-  inst->GetResultsTypes()[1] = Type{DataType::INT64, type0.GetDimSizes()};
+  inst->GetResultsTypes()[1] = Type{DataType::INT32, type0.GetDimSizes()};
 }
 
 bool TypeLegalizer::RunOnBasicBlock(BasicBlock* bb) {
